requirements.txt
openai==0.27.4
pinecone-client==2.2.1
python-dotenv
Gunicorn
Flask
requests
requests_oauthlib
tweepy
langchain==0.0.195
redis
numpy
pandas
gdown
typing
tiktoken
pytz
weaviate-client
urllib3<2.0.0
